backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
click==7.1.2
Django==3.2.3
django-cors-headers==3.7.0
django-extensions==3.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.4
regex==2020.11.13
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
